+++ src/main/java/kitchenpos/ui/menu/response/MenuGroupResponse.java
@@ -1,6 +1,6 @@
 package kitchenpos.ui.menu.response;
 
-import kitchenpos.menu.MenuGroup;
+import kitchenpos.menugroup.MenuGroup;
 
 public class MenuGroupResponse {
