+++ src/test/java/kitchenpos/ServiceTestHelper.java
@@ -3,7 +3,7 @@ package kitchenpos;
 import kitchenpos.menu.MenuGroupService;
 import kitchenpos.menu.MenuService;
 import kitchenpos.menu.request.MenuCreateRequest;
-import kitchenpos.menu.request.MenuGroupCreateRequest;
+import kitchenpos.menugroup.request.MenuGroupCreateRequest;
 import kitchenpos.menu.request.MenuProductRequest;
 import kitchenpos.order.OrderService;
 import kitchenpos.table.TableService;
@@ -18,7 +18,7 @@ import kitchenpos.table.TableGroupService;
 import kitchenpos.table.request.OrderTableIdRequest;
 import kitchenpos.table.request.TableGroupCreateRequest;
 import kitchenpos.menu.Menu;
-import kitchenpos.menu.MenuGroup;
+import kitchenpos.menugroup.MenuGroup;
 import kitchenpos.menu.MenuProductRepository;
 import kitchenpos.order.Order;
 import kitchenpos.order.OrderStatus;
