+++ src/test/java/kitchenpos/menugroup/MenuGroupServiceTest.java
@@ -1,6 +1,7 @@
-package kitchenpos.menu;
+package kitchenpos.menugroup;
 
 import kitchenpos.ServiceTestHelper;
+import kitchenpos.menugroup.MenuGroup;
 import org.junit.jupiter.api.DisplayNameGeneration;
 import org.junit.jupiter.api.DisplayNameGenerator;
 import org.junit.jupiter.api.Test;
