+++ src/test/java/kitchenpos/ui/menu/MenuAcceptanceTestUtil.java
@@ -6,11 +6,11 @@ import io.restassured.response.ExtractableResponse;
 import io.restassured.response.Response;
 import kitchenpos.menu.MenuGroupService;
 import kitchenpos.menu.request.MenuCreateRequest;
-import kitchenpos.menu.request.MenuGroupCreateRequest;
+import kitchenpos.menugroup.request.MenuGroupCreateRequest;
 import kitchenpos.menu.request.MenuProductRequest;
 import kitchenpos.product.ProductService;
 import kitchenpos.product.request.ProductCreateRequest;
-import kitchenpos.menu.MenuGroup;
+import kitchenpos.menugroup.MenuGroup;
 import kitchenpos.product.Product;
 import kitchenpos.ui.ControllerAcceptanceTestHelper;
 import kitchenpos.ui.menu.response.MenuResponse;
