+++ src/test/java/kitchenpos/ui/order/OrderAcceptanceTestUtil.java
@@ -7,7 +7,7 @@ import io.restassured.response.Response;
 import kitchenpos.menu.MenuGroupService;
 import kitchenpos.menu.MenuService;
 import kitchenpos.menu.request.MenuCreateRequest;
-import kitchenpos.menu.request.MenuGroupCreateRequest;
+import kitchenpos.menugroup.request.MenuGroupCreateRequest;
 import kitchenpos.menu.request.MenuProductRequest;
 import kitchenpos.table.TableService;
 import kitchenpos.order.request.OrderCreateRequest;
@@ -17,7 +17,7 @@ import kitchenpos.order.request.TableCreateRequest;
 import kitchenpos.product.ProductService;
 import kitchenpos.product.request.ProductCreateRequest;
 import kitchenpos.menu.Menu;
-import kitchenpos.menu.MenuGroup;
+import kitchenpos.menugroup.MenuGroup;
 import kitchenpos.order.OrderStatus;
 import kitchenpos.table.OrderTable;
 import kitchenpos.product.Product;
